moved Databse class to own file; Classes now use namespaces; Updated type hints with namespace;
